[refactor] user RedisUtil 위치 변경 1.RedisUtil 위치 login -> common
