+++ easytask/src/main/java/com/easytask/easytask/common/response/BaseResponseStatus.java
@@ -22,7 +22,7 @@ public enum BaseResponseStatus {
     NOT_VALID_EMAIL(false, HttpStatus.BAD_REQUEST.value(),"이메일 형식에 맞지 않습니다"),
     NOT_FIND_USER(false,HttpStatus.NOT_FOUND.value(),"이메일 또는 비밀번호를 확인해주세요"),
     INVALID_REQUEST(false,HttpStatus.BAD_REQUEST.value(), "잘못된 요청입니다."),
-    INVALID_TOKEN(false,HttpStatus.BAD_REQUEST.value(), "토큰이 잘못되었습니다"),
+    INVALID_TOKEN(false,HttpStatus.BAD_REQUEST.value(), "유효하지 않은 토큰입니다."),
