+++ easytask/src/main/java/com/easytask/easytask/src/user/entity/PossibleTask.java
@@ -30,9 +30,4 @@ public class PossibleTask{
         this.categoryBig = categoryBig;
         this.categorySmall = categorySmall;
     }
-    public void updateSkill(String categoryBig, String categorySmall){
-        this.categoryBig=categoryBig;
-        this.categorySmall=categorySmall;
-    }
-
 }
