+++ easytask/src/main/java/com/easytask/easytask/src/user/entity/TaskAbility.java
@@ -30,8 +30,4 @@ public class TaskAbility{
         this.categoryBig = categoryBig;
         this.categorySmall = categorySmall;
     }
-    public void updateSkill(String categoryBig, String categorySmall){
-        this.categoryBig=categoryBig;
-        this.categorySmall=categorySmall;
-    }
 }
